% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{createRainbowColors}
\alias{createRainbowColors}
\title{createRainbowColors
Create a list of colors for each value v}
\usage{
createRainbowColors(v)
}
\arguments{
\item{v}{a vector of characters}
}
\value{
a list of value=color
}
\description{
createRainbowColors
Create a list of colors for each value v
}
