\name{hostline}
\alias{hostline}
\docType{data}
\title{
  A host infomation over time
}
\description{
  kind of host time line
}
\usage{data("hostline")}
\format{
  A data frame with 8 observations on the following 5 variables.
  \describe{
    \item{\code{level}}{a character vector}
    \item{\code{label}}{a character vector}
    \item{\code{ID}}{a character vector}
    \item{\code{timestart}}{a character vector}
    \item{\code{timeend}}{a character vector}
  }
}
\examples{
data(hostline)
print(hostline)
}
\keyword{datasets}
