% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapttProxy.R
\name{mapttProxy}
\alias{mapttProxy}
\title{mapttProxy}
\usage{
mapttProxy(mapttId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{mapttId}{widget instance identifier}

\item{session}{shiny session}
}
\description{
get mapttProxy
}
\examples{
\dontrun{
library(SMITIDvisu)
## server.R
mapttProxy <- mapttProxyProxy("mapttOutput")
}

}
