% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapttProxy.R
\name{mapttSelectHost}
\alias{mapttSelectHost}
\title{mapttSelectHost}
\usage{
mapttSelectHost(mapttProxy, hostId)
}
\arguments{
\item{mapttProxy}{mapttProxy instance}

\item{hostId}{the id of the host to select}
}
\description{
select a host on a MapTT instance
}
\examples{
\dontrun{
library(SMITIDvisu)
data(transmissiontree)
## server.R
mapttProxy("mapttOutput") \%>\% mapttSelectHost()
}
}
\seealso{
\code{\link{maptt}}
}
