% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mstCompute}
\alias{mstCompute}
\title{compute the minimum spanning tree}
\usage{
mstCompute(mat)
}
\arguments{
\item{mat}{weighted matrix representing nodes connection (edges weight)}
}
\value{
a matrix with 1 if nodes are linked, 0 otherwise.
}
\description{
compute the minimum spanning tree of a matrix representing edges between nodes (of a graph)
}
