% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstVariantProxy.R
\name{mstVariantProxy}
\alias{mstVariantProxy}
\title{mstVariantProxy}
\usage{
mstVariantProxy(mstVid, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{mstVid}{widget instance identifier}

\item{session}{shiny session}
}
\description{
get mstVariantProxy
}
\examples{
\dontrun{
library(SMITIDvisu)
## server.R
mstVariantProxy <- mstVaraintProxy("mstvariantoutput")
}

}
