% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMITIDvisu.R
\docType{package}
\encoding{UTF-8}
\name{SMITIDvisu-package}
\alias{SMITIDvisu-package}
\alias{SMITIDvisu}
\title{Visualize Data for Host and Viral Population from SMITIDstruct using HTMLwidgets}
\description{
Visualisation tools for SMITIDstruct package.
Allow to visualize host timeline, transmission tree, index diversities and variant graph using HTMLwidgets.
It mainly using D3JS, noUiSlider and FileSaver javascript libraries.
}
\details{
\tabular{ll}{
         Package: \tab SMITIDvisu\cr
         Type: \tab Package\cr
         Version: \tab 0.0.8\cr
         Date: \tab 2020-11-04\cr
         License: \tab GPL (>=3)\cr
         }
}
\examples{
\dontrun{
 library(SMITIDvisu)
 demo.SMITIDvisu.run()
}
}
\author{
Jean-Francois Rey \email{jean-francois.rey@inrae.fr}

Julien Boge \email{julien.boge.u@gmail.com}
}
