\name{st.dist113_2}
\alias{st.dist113_2}
\docType{data}
\title{
  Distance matrix of observed variants sequences of a host 113 at time 2 from simulation.
}
\description{
  Levenshtein Distance matrix with rows and cols label as sequences ID.
}
\usage{data("st")}
\format{
  The format is:
 num [1:23, 1:23] 0 1 1 1 2 1 1 1 1 2 ...
}
\examples{
data(st)
}
\keyword{datasets}
