\name{st.dist113_all}
\alias{st.dist113_all}
\docType{data}
\title{
  Distance matrix of observed variants sequences of a host 113 at time 2, 3 and 4 from simulation.
}
\description{
  Levenshtein Distance matrix with rows and cols label as sequences ID.
  Unique sequence variants observed on host 113 at time 2, 3 and 4 from a simulation.
}
\usage{data("st")}
\format{
  The format is:
 num [1:51, 1:51] 0 1 1 1 1 1 3 1 1 1 ...
}
\examples{
data(st)
}
\keyword{datasets}
