\name{st.listTimeProp113}
\alias{st.listTimeProp113}
\docType{data}
\title{
  List of variants ID with subvector for time and value.
}
\description{
  A list indexed by variants sequences ID. Each element contain a time and value vector for time of observation and proportions observed at this time.
}
\usage{data("st")}
\examples{
data(st)
}
\keyword{datasets}
