\name{st.prop113_all}
\alias{st.prop113_all}
\docType{data}
\title{
  Variants proportions and count for an host 113 at time 2, 3 and 4 from simulation.
}
\description{
  A data.frame with label "ID", "proportion" and "count" for an host 113 at time 2, 3 and 4 from simulation.
    Each row is a sequence.
}
\usage{data("st")}
\format{
  A data frame with 51 observations on the following 3 variables.
  \describe{
    \item{\code{ID}}{a character vector}
    \item{\code{proportion}}{a numeric vector}
    \item{\code{count}}{a numeric vector}
  }
}
\examples{
data(st)
}
\keyword{datasets}
