% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeLineProxy.R
\name{timeLineProxy}
\alias{timeLineProxy}
\title{timeLineProxy
get an instance of a timeline}
\usage{
timeLineProxy(tlid, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{tlid}{a timeline instance id}

\item{session}{shiny session}
}
\value{
an object of class timeline_proxy
}
\description{
timeLineProxy
get an instance of a timeline
}
\examples{
\dontrun{
## server.R
## output server variable
output$timeline <- renderTimeLine({
         timeLine(data.frame(), "")
       })
## ui.R
timeLineOutput("timeline")
## server.R
tlproxy <- timeLineProxy("timeline")
}
}
