% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeLineProxy.R
\name{updateTimeLine}
\alias{updateTimeLine}
\title{updateTimeLine}
\usage{
updateTimeLine(tlProxy, data, title)
}
\arguments{
\item{tlProxy}{a timeline proxy instance}

\item{data}{new data}

\item{title}{new title}
}
\description{
updateTimeLine
}
\examples{
\dontrun{
## server.R
## output server variable
output$timeline <- renderTimeLine({
         timeLine(data.frame(), "")
       })
## ui.R
timeLineOutput("timeline")
## server.R
timeLineProxy("timeline") \%>\% updateTimeLine(newtimeline, "newId")
}
}
\seealso{
\code{\link{timeLine}}
}
