% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmissionTreeProxy.R
\name{updateTransmissionTree}
\alias{updateTransmissionTree}
\title{updateTransmissionTree}
\usage{
updateTransmissionTree(TTProxy, nodes, edges, options = NULL)
}
\arguments{
\item{TTProxy}{transmissionTreeProxy instance}

\item{nodes}{a data.frame that represent hosts status in time with ID, status and time in columns}

\item{edges}{a data.frame that represent tramsmission link between hosts (pathogens) with ID, source, weight, target and time in columns}

\item{options}{transmissionTree new options}
}
\description{
update (redraw) an instance of a transmissionTree
}
\examples{
\dontrun{
library(SMITIDvisu)
data(transmissionTree)
## server.R
transmissionTreeProxy("transmissionTreeoutput") \%>\% updatetransmissionTree(tt.nodes,tt.edges)
}
}
\seealso{
\code{\link{transmissionTree}}
}
